Change size_t to unsigned long int to solve problem compatibility between 32bits and 64 bits when printing with %lu

+++ thesquid.c
@@ -2888,7 +2888,7 @@ void SquadBenchmark(
     fprintf(stream, "%s", header);
 
     // Loop on the payload
-    for (size_t sizePayload = 9; 
+    for (unsigned long int sizePayload = 9; 
       sizePayload <= maxSizePayload; sizePayload *= 10) {
 
       // Create a dummy buffer with the size of the payload
